backend/pyproject.toml
[tool.poetry]
name = "pranita"
version = "0.1.0"
description = ""
authors = ["PranitaR12 <prandive12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
pillow = "^10.0.1"
django-cors-headers = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
